# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/products/{id}/media": {
      "get": {
        "description": "Tartiblangan. O‘chirilgan mahsulot uchun ham ishlaydi (ochiq `GET /products/:slug` dan farqli).",
        "operationId": "ProductMediaAdminController_findAll",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mahsulot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media fayllar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductMediaAdminResponseDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mahsulot media ro‘yxati",
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "description": "`multipart/form-data`: `file` + ixtiyoriy `type`.\n\n• `IMAGE`, `IMAGE_360` — JPG, PNG, WEBP\n• `VIDEO_360` — MP4\n\nMaksimal hajm — 10 MB. Tur fayl MAZMUNIDAN aniqlanadi: kengaytmani o‘zgartirish (`rasm.svg` → `rasm.jpg`) yordam bermaydi.\n\nYangi fayl ro‘yxat oxiriga qo‘shiladi.",
        "operationId": "ProductMediaAdminController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mahsulot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadProductMediaBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Yuklandi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProductMediaAdminResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Fayl yo‘q yoki turi mos emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "413": {
            "description": "Fayl 10 MB dan katta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Media yuklash",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/products/{id}/media/order": {
      "patch": {
        "description": "`mediaIds` — mahsulotning BARCHA media ID lari yangi tartibda. Qisman ro‘yxat yoki begona ID — 400.",
        "operationId": "ProductMediaAdminController_reorder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mahsulot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderProductMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Yangi tartibdagi media",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductMediaAdminResponseDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ro‘yxat to‘liq emas yoki begona ID bor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Media tartibini o‘zgartirish",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/media/{id}": {
      "delete": {
        "description": "Bazadagi yozuv va diskdagi fayl o‘chiriladi.",
        "operationId": "ProductMediaAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Media ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "O‘chirilgan media",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProductMediaAdminResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Media topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Media o‘chirish",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/calculator/quote": {
      "post": {
        "description": "Paddon soni bo‘yicha m², og‘irlik, summa; `regionId` + `transportTypeId` berilsa — nechta transport va yo‘l kira.\n\n🔒 Narx va filial so‘rovdan OLINMAYDI: narx — mijoz filialining narxi, shaxsiy chegirmasi bilan; filial — tokendan.\n\nBuyurtma berilganda summa AYNAN shu yo‘l bilan qayta hisoblanadi.",
        "operationId": "CalculatorController_quote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hisob natijasi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/QuoteResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Paddon soni noto‘g‘ri, mahsulot takrorlangan, viloyat/transport faqat bittasi berilgan yoki summa juda katta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Token yo‘q, muddati o‘tgan yoki hisob faol emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Optom mijoz tokeni emas yoki vaqtinchalik parol almashtirilmagan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot filialda sotilmaydi yoki bu yo‘nalishga tarif yo‘q",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "To‘liq hisob: mahsulotlar + yo‘l kira",
        "tags": [
          "Calculator"
        ]
      }
    },
...
    "/api/v1/orders": {
      "post": {
        "description": "Mahsulotlar (paddon bilan), ixtiyoriy yetkazib berish (viloyat + transport), to‘lov usuli.\n\n🔒 Summa backendda QAYTA hisoblanadi — `POST /calculator/quote` bilan aynan bir xil yo‘l. Frontend yuborgan narx/summa e’tiborga olinmaydi.\n\nOmborda yetarli bo‘lmasa — 409 (qancha borligi aytilmaydi).\n\nNatija: `NEW` holatdagi buyurtma va `PENDING` to‘lov yozuvi.",
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Buyurtma yaratildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/OrderCustomerResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Paddon/mahsulot noto‘g‘ri, viloyat va transport birga emas, koordinatalar noto‘liq yoki olib ketishda berilgan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Token yo‘q, muddati o‘tgan yoki hisob faol emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Optom mijoz tokeni emas yoki vaqtinchalik parol almashtirilmagan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot filialda sotilmaydi yoki tarif yo‘q",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Omborda yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buyurtma berish",
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vodiy Kafel API",
    "description": "Vodiy Kafel savdo platformasining backend API'si.\n\n### Javob formati\nBarcha muvaffaqiyatli javoblar `{ \"data\": ... }` ichida keladi.\nSahifalangan ro'yxatlarda qo'shimcha `meta` bo'ladi.\n\nBarcha xatolar bir xil formatda — `ApiErrorDto` modeliga qarang.\n\n### Autentifikatsiya\nToken `Authorization: Bearer <token>` sarlavhasida yuboriladi.\nYuqoridagi **Authorize** tugmasi orqali kiritishingiz mumkin.\n\n### Muhim qoidalar\n- Mahsulot javoblarida ombordagi **aniq son yo'q** — faqat `stockStatus`.\n- Buyurtma summasini backend o'zi hisoblaydi; frontend yuborgan narx e'tiborga olinmaydi.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status kodi"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Status nomi"
          },
          "message": {
            "description": "Xato matni. Validatsiya xatosida — matnlar massivi, aks holda bitta matn.",
            "oneOf": [
              {
                "type": "string",
                "example": "Mahsulot topilmadi"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "pallets must not be less than 1"
                ]
              }
            ]
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products",
            "description": "So‘rov manzili"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-08T17:02:11.419Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "1339b1a6-5cf7-4d0f-9d4b-10bfbd295f49",
            "description": "Kuzatuv ID. Xato haqida xabar berganda shu ID ni ayting — log‘dan aniq so‘rov topiladi."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "ProductMediaAdminResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmtz0a1b2c3d4e5f6g7h8i9j"
          },
          "productId": {
            "type": "string",
            "example": "cmtz0a1b2c3d4e5f6g7h8i9j"
          },
          "url": {
            "type": "string",
            "example": "/uploads/products/0b6f1c1e-6a4f-4a57-9a3b-2f7c1d9e8a11.jpg"
          },
          "type": {
            "type": "string",
            "enum": [
              "IMAGE",
              "IMAGE_360",
              "VIDEO_360"
            ],
            "example": "IMAGE"
          },
          "sortOrder": {
            "type": "number",
            "description": "Ko‘rsatish tartibi (0 — birinchi)",
            "example": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-14T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "productId",
          "url",
          "type",
          "sortOrder",
          "createdAt"
        ]
      },
      "UploadProductMediaBodyDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "IMAGE",
              "IMAGE_360",
              "VIDEO_360"
            ],
            "default": "IMAGE",
            "description": "`IMAGE` va `IMAGE_360` — JPG/PNG/WEBP surat, `VIDEO_360` — MP4 video."
          },
          "file": {
            "type": "string",
            "format": "binary",
            "description": "Fayl: JPG, PNG, WEBP yoki MP4, 10 MB gacha"
          }
        },
        "required": [
          "file"
        ]
      },
      "ReorderProductMediaDto": {
        "type": "object",
        "properties": {
          "mediaIds": {
            "description": "Mahsulotning BARCHA media ID lari — yangi tartibda. Birinchisi kartada ko‘rinadi (agar u `IMAGE` bo‘lsa).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "mediaIds"
        ]
      },
...
      "QuoteResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteItemResponseDto"
            }
          },
          "totalPallets": {
            "type": "number",
            "example": 15
          },
          "totalSqm": {
            "type": "string",
            "example": "19.8"
          },
          "totalWeightKg": {
            "type": "string",
            "example": "445"
          },
          "itemsTotal": {
            "type": "string",
            "example": "1558800"
          },
          "transport": {
            "nullable": true,
            "description": "Yetkazib berishsiz hisobda — `null`",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryResponseDto"
              }
            ]
          },
          "deliveryTotal": {
            "type": "string",
            "example": "12000000"
          },
          "grandTotal": {
            "type": "string",
            "example": "13558800"
          }
        },
        "required": [
          "items",
          "totalPallets",
          "totalSqm",
          "totalWeightKg",
          "itemsTotal",
          "deliveryTotal",
          "grandTotal"
        ]
      },
      "QuoteItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "Mahsulot ID"
          },
          "pallets": {
            "type": "number",
            "description": "Paddon soni (kv² emas — TZ 3.3)",
            "minimum": 1,
            "maximum": 100000,
            "example": 10
          }
        },
        "required": [
          "productId",
          "pallets"
        ]
      },
      "QuoteRequestDto": {
        "type": "object",
        "properties": {
          "items": {
            "minItems": 1,
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteItemDto"
            }
          },
          "regionId": {
            "type": "string",
            "description": "Yetkazib berish viloyati. `transportTypeId` bilan BIRGA."
          },
          "transportTypeId": {
            "type": "string",
            "description": "Transport turi. `regionId` bilan BIRGA. Ikkalasi ham berilmasa — yetkazib berishsiz hisob."
          }
        },
        "required": [
          "items"
        ]
      },
...
      "OrderCustomerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmtz0a1b2c3d4e5f6g7h8i9j"
          },
          "orderNumber": {
            "type": "string",
            "example": "VK-2026-000001"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "SEARCHING_TRANSPORT",
              "LOADING",
              "DELIVERING",
              "DELIVERED",
              "CANCELLED"
            ],
            "example": "NEW"
          },
          "source": {
            "type": "string",
            "enum": [
              "WEBSITE",
              "TELEGRAM",
              "PHONE",
              "ADMIN"
            ],
            "example": "WEBSITE"
          },
          "branchName": {
            "type": "string",
            "example": "Vodiy Kafel — Farg‘ona"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "totalPallets": {
            "type": "number",
            "example": 10
          },
          "totalSqm": {
            "type": "string",
            "example": "14.4"
          },
          "totalWeightKg": {
            "type": "string",
            "example": "325"
          },
          "itemsTotal": {
            "type": "string",
            "example": "1123200"
          },
          "deliveryTotal": {
            "type": "string",
            "example": "0"
          },
          "grandTotal": {
            "type": "string",
            "example": "1123200"
          },
          "delivery": {
            "nullable": true,
            "description": "Olib ketishda — `null`",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderDeliveryDto"
              }
            ]
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderPaymentDto"
            }
          },
          "statusHistory": {
            "description": "Holatlar tarixi — vaqt belgisi bilan (TZ 3.4)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderStatusEntryDto"
            }
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-14T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "source",
          "branchName",
          "items",
          "totalPallets",
          "totalSqm",
          "totalWeightKg",
          "itemsTotal",
          "deliveryTotal",
          "grandTotal",
          "payments",
          "statusHistory",
          "createdAt"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "minItems": 1,
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteItemDto"
            }
          },
          "regionId": {
            "type": "string",
            "description": "Yetkazib berish viloyati — `transportTypeId` bilan BIRGA. Ikkalasi ham berilmasa — olib ketish (yo‘l kira 0)."
          },
          "transportTypeId": {
            "type": "string",
            "description": "Transport turi — `regionId` bilan BIRGA"
          },
          "exactLat": {
            "type": "number",
            "description": "Xaritada belgilangan aniq nuqta (TZ 3.13) — `exactLng` bilan birga, faqat yetkazib berishda. ⚠ Narxga TA’SIR QILMAYDI — faqat logistika.",
            "minimum": -90,
            "maximum": 90,
            "example": 41.311081
          },
          "exactLng": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "example": 69.240562
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "CASH",
              "CARD",
              "BANK_TRANSFER"
            ],
            "description": "`CASH` — naqd, `CARD` — karta orqali onlayn, `BANK_TRANSFER` — shartnoma bo‘yicha o‘tkazma"
          },
          "note": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": [
          "items",
          "paymentMethod"
        ]
      },
...
    }
  }
}